adjust file names
Rename the login.html file to index.html; rename the original index.html file to lower_index.html, and synchronously modify the code in other interfaces that jumps to these pages.  Keep consistency with the names of other types of jump pages, that is, the first page for all jumps is index.html.